+++ src/api/setAccessToken.ts
@@ -6,8 +6,6 @@ interface SetAccessToken {
     token: string;
 };
 
-function setAccessToken({ client, token }: SetAccessToken) {
+export function setAccessToken({ client, token }: SetAccessToken) {
     return client.setUserState(OAUTH2_ACCESS_TOKEN_PATH, token, {backend: true});
 };
-
-export default setAccessToken;
